pyproject.toml
[project]
name = "bfi-2-j-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bootstrap-flask>=2.4.1",
    "flask-wtf>=1.2.2",
    "pandas>=2.2.3",
]
